ColorSchemeDialog options all properly set
Now, if you select a light theme or dark theme, or a high contrast theme, from the Color Scheme Dialog, the setting is now set properly and remembered. And it is also backwards-compatible with previous versions, as earlier versions will default to a colored theme instead.

Merged the "Light theme" and "Dark theme" options, into a new "Basic theme" section. This should also help simplify the dialog a decent amount. Also resized the dialog to make it a bit shorter.

Also removed the "High Contrast" option in the View menu, as the new dialog will set the theme properly.

+++ PathfinderJson/App.xaml.cs
@@ -261,28 +261,31 @@ namespace PathfinderJson
             string appDataPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), "PathfinderJson");
             string errorLogPath = Path.Combine(appDataPath, "ErrorLogs");
 
-            Exception ex = e.Exception;
-
-            StringBuilder sb = new StringBuilder();
-
-            sb.AppendLine(DateTime.UtcNow.ToString("yyyyMMddTHH:mm:ssZ"));
-            sb.AppendLine("VERSION " + AppVersion.ToString());
-            sb.AppendLine("--------------------------------------");
-            sb.AppendLine(ex.GetType().FullName);
-            sb.AppendLine(ex.Message);
-            sb.AppendLine(ex.ToString());
-
-            if (ex.InnerException != null)
+            if (Directory.Exists(errorLogPath))
             {
-                sb.AppendLine("INNER EXCEPTION:");
-                sb.AppendLine(ex.InnerException.GetType().FullName);
-                sb.AppendLine(ex.InnerException.Message);
-                sb.AppendLine(ex.InnerException.ToString());
+                Exception ex = e.Exception;
+
+                StringBuilder sb = new StringBuilder();
+
+                sb.AppendLine(DateTime.UtcNow.ToString("yyyyMMddTHH:mm:ssZ"));
+                sb.AppendLine("VERSION " + AppVersion.ToString());
+                sb.AppendLine("--------------------------------------");
+                sb.AppendLine(ex.GetType().FullName);
+                sb.AppendLine(ex.Message);
+                sb.AppendLine(ex.ToString());
+
+                if (ex.InnerException != null)
+                {
+                    sb.AppendLine("INNER EXCEPTION:");
+                    sb.AppendLine(ex.InnerException.GetType().FullName);
+                    sb.AppendLine(ex.InnerException.Message);
+                    sb.AppendLine(ex.InnerException.ToString());
+                }
+
+                sb.AppendLine("END FILE");
+
+                await File.WriteAllTextAsync(Path.Combine(errorLogPath, DateTime.UtcNow.ToString("yyyyMMddTHHmmssZ") + ".txt"), sb.ToString(), Encoding.UTF8);
             }
-
-            sb.AppendLine("END FILE");
-
-            await File.WriteAllTextAsync(Path.Combine(errorLogPath, DateTime.UtcNow.ToString("yyyyMMddTHHmmssZ") + ".txt"), sb.ToString(), Encoding.UTF8);
 
             MessageBox.Show("An error has occurred and PathfinderJSON may not be able to continue.\n\n" +
                 "An error log file was created.\n\n" +
